Conflicts:
	.gitignore
	Messages/Msg.cpp
	Messages/Msg.h
	Misc/DHCPConst.h
	Misc/DHCPServer.cpp
	Misc/Portable.h
	Misc/Portable.h.in
	Port-bsd/lowlevel-bsd.c
	Port-linux/daemon.cpp
	RelOptions/RelOptRemoteID.cpp
	Requestor/Makefile.in
	Requestor/ReqTransMgr.cpp
	Requestor/Requestor.cpp
	SrvCfgMgr/NodeClientSpecific.cpp
	SrvCfgMgr/SrvCfgIface.cpp
	SrvCfgMgr/SrvCfgMgr.cpp
	SrvIfaceMgr/SrvIfaceMgr.cpp
	SrvMessages/SrvMsg.cpp
	SrvMessages/SrvMsgLeaseQueryReply.cpp
	SrvTransMgr/SrvTransMgr.cpp

+++ ClntCfgMgr/ClntParsAddrOpt.cpp
@@ -16,8 +16,8 @@
 
 TClntParsAddrOpt::TClntParsAddrOpt()
 {
-    Pref=ULONG_MAX;
-    Valid=ULONG_MAX;
+    Pref=UINT_MAX;
+    Valid=UINT_MAX;
 }
 
 long TClntParsAddrOpt::getPref()
